Decoupled the DiagramIndex
