+++ src/main/java/org/reactome/server/tools/diagram/exporter/raster/gif/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Tools to generate a GIF. See {@link org.reactome.server.tools.diagram.exporter.raster.gif.AnimatedGifEncoder}
+ * for more information.
+ */
+package org.reactome.server.tools.diagram.exporter.raster.gif;
